more info in log for rule executions -- still need to get the order of predicate execution in there with the #of statements considered for the predicate.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@307 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/store/DataLoader.java
@@ -449,7 +449,7 @@ public class DataLoader {
                 log.info
                     ( e.getStatementsProcessed() + 
                       " stmts added in " + 
-                      ((double)e.getTimeElapsed()) / 1000d +
+                      (e.getTimeElapsed() / 1000d) +
                       " secs, rate= " + 
                       e.getInsertRate() 
                       );
@@ -484,6 +484,10 @@ public class DataLoader {
                 
                 /*
                  * compute the closure.
+                 * 
+                 * FIXME closure stats are not being reported out, e.g., to the DataLoader.
+                 * 
+                 * Also, batch closure logically belongs in the outer method.
                  */
                 
                 log.info("Computing closure.");
